fix(figma-plugin): compare variable floats in single precision
Figma stores variable floats in float32, so 0.7 reads back as
0.699999988079071. `valuesEqual` compared numbers with Object.is on float64,
which rejected every value not representable in single precision and failed
the post-sync verification with "variable shape/opacity/dim value does not
match" — even though all 699 variables had been written correctly.

Latent until now: the previously published 597 variables were all colors and
integers. Phase 2 introduced the first fractional scalars (opacity 0.35/0.7,
leading 1.2), and shape/opacity/dim is simply the first one alphabetically.
type/leading/tight would have failed next.

Colors already avoided this — they compare per channel through closeTo() with
a tolerance — but the scalar path never got the same treatment.

Compares via Math.fround rather than an epsilon: it models exactly what Figma
does, stays an exact comparison, and does not lose precision on large values
the way closeTo's absolute 1e-6 would. Verified it accepts the stored values
(0.7, 0.35, 1.2) while still rejecting genuine mismatches (0.7 vs 0.8).

+++ figma/plugins/collider-token-sync/code.ts
@@ -261,7 +261,11 @@ type PluginVariableValue =
 
 function valuesEqual(expected: unknown, actual: unknown) {
   if (typeof expected === 'number' && typeof actual === 'number') {
-    return Object.is(expected, actual);
+    // Figma stores variable floats in single precision, so 0.7 reads back as
+    // 0.699999988079071. Comparing float64 exactly would reject every value
+    // that is not representable in float32; compare in float32 instead.
+    // Colors already sidestep this via closeTo() on each channel.
+    return Object.is(Math.fround(expected), Math.fround(actual));
   }
 
   if (typeof expected === 'string' && typeof actual === 'string') {
